Proceeding

+++ src/controllers/videoControllers.js
@@ -1,4 +1,3 @@
-import { redirect, render } from 'express/lib/response';
 import Video from "../models/Video";
 import User from "../models/User";
 
@@ -30,34 +29,50 @@ const videos = [
 ];
 
 export const home = async (req, res) => {
-    const videos = await Video.find({}).sort();
+    // 모든 controllers에는 id가 붙는다. video의 id도 가져온다.
+    // sort는 정렬이다. sort를 사용해서 video를 순서대로 home에 보낸다. 그러면 video는 순서대로 노출된다.
+    const videos = await Video.find({}).sort({createdAt:"desc"})
+    .populate("owner");
     return res.render("home", { pageTitle: "Home", videos});
 };
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
-    const owner = await User.findById(video.owner);
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
     if(video === null){   
         return res.render("404", {pageTitle: 'video not found.'});
     }
-        return res.render("watch", { pageTitle: video.title, video, owner });
+        return res.render("watch", video);
 };
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id }} = req.session;
     const video = await Video.findById(id);
     //${video.title}`
     if (!video) {
         return res.status(404).render("404", {pageTitle: "video not found."});        
     }
+    // video.owner은 obj  _id는 string
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
         return res.render("edit", {pageTitle: `Editing: ${video.title}`, video});
 };
 
 export const postEdit = async (req, res) =>{
+    const {
+        user: { _id },
+    } = req.session;
     const { id } = req.params;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({ _id: id });
+    const video = await Video.findById(id); 
+    // exists({ _id: id })
     if (!video) {
         return res.render("404", { pageTitle: "Video not found."});
+    }
+    // video.owner은 obj  _id는 string
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -79,7 +94,7 @@ export const postUpload = async (req, res) =>{
     const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
@@ -87,6 +102,9 @@ export const postUpload = async (req, res) =>{
             owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (err) {
         console.log(err);
@@ -96,8 +114,19 @@ export const postUpload = async (req, res) =>{
 
 export const deleteVideo = async (req, res) =>{
     const { id } = req.params;
-    console.log(id);
+    const {
+        user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id), 1);
     return res.redirect("/");
 }
 
@@ -108,9 +137,9 @@ export const search = async (req, res) => {
         // keyword가 있다면 videos=[]을 업데이트
         videos = await Video.find({
             title: {
-                $regex: new RegExp(keyword, "i")
+                $regex: new RegExp(keyword, "i"),
             },
-        });
+        }).populate("owner");
     }
     return res.render("search", { pageTitle: "Search", videos });
 }
